+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/js/MessageBridgeRatDispatcherSpec.kt
@@ -33,7 +33,8 @@ class MessageBridgeRatDispatcherSpec {
         ActionType.SEND_MESSAGE_TO_CONTACT.action to Actype.SEND_MESSAGE_TO_CONTACT,
         ActionType.SEND_MESSAGE_TO_CONTACT_ID.action to Actype.SEND_MESSAGE_TO_CONTACT_ID,
         ActionType.SEND_MESSAGE_TO_MULTIPLE_CONTACTS.action to Actype.SEND_MESSAGE_TO_MULTIPLE_CONTACTS,
-        ActionType.GET_POINTS.action to Actype.GET_POINTS
+        ActionType.GET_POINTS.action to Actype.GET_POINTS,
+        ActionType.GET_HOST_ENVIRONMENT_INFO.action to Actype.GET_HOST_ENVIRONMENT_INFO
     )
 
     @Test