+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 ## CHANGELOG
 
-### 3.9.0 (2021-XX-XX)
+### 3.9.1 (2022-01-20)
+
+**SDK**
+- **Fix:** Unable to launch a mini app while the device is offline
+
+### 3.9.0 (2021-12-17)
+**SDK**
+- **Feature:** Added `languageCode` parameter in `MiniApp.getMiniAppManifest` to support for internationalized manifest.
+- **Feature:** Added `hostLocale` in `HostEnvironmentInfo` to provide default language value from Host App.
+- **Feature:** Added `promotionalImageUrl` and `promotionalText` in MiniAppInfo model.
+- **Feature:** Added support for taking picture from camera in Mini App.
+
 **Sample App**
 - **Feature:** Added production and staging toggle to change environments.
+- **Feature:** Adding Sharing option to display the promotional content via `MiniAppDisplay`.
 
 ### 3.8.0 (2021-11-09)
 **SDK**